adding subtraction test + pass into generic class the common data
