+++ src/test/java/calculator/GenericTestData.java
@@ -1,0 +1,20 @@
+package calculator;
+
+import com.epam.tat.module4.Calculator;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
+
+public class GenericTestData {
+
+    protected Calculator calculator;
+
+    @BeforeClass
+    public void setUp() {
+        calculator = new Calculator();
+    }
+
+    @AfterClass
+    public void tearDown() {
+        calculator = null;
+    }
+}
